Move stuff in/to src/Elewant/Herding/Application
