+++ src/Elewant/Herding/Application/Commands/AbandonHerd.php
@@ -2,14 +2,15 @@
 
 declare(strict_types=1);
 
-namespace Elewant\Herding\Model\Commands;
+namespace Elewant\Herding\Application\Commands;
 
-use Elewant\Herding\Model\HerdId;
-use Elewant\Herding\Model\ShepherdId;
+use Elewant\Herding\DomainModel\Herd\HerdId;
+use Elewant\Herding\DomainModel\ShepherdId;
+use Elewant\Herding\DomainModel\SorryThatIsAnInvalid;
 use Prooph\Common\Messaging\Command;
 use Prooph\Common\Messaging\PayloadTrait;
 
-class AbandonHerd extends Command
+final class AbandonHerd extends Command
 {
     use PayloadTrait;
 
@@ -18,11 +19,17 @@ class AbandonHerd extends Command
         return new self(['herdId' => $herdId, 'shepherdId' => $shepherdId]);
     }
 
+    /**
+     * @throws SorryThatIsAnInvalid
+     */
     public function herdId(): HerdId
     {
         return HerdId::fromString($this->payload['herdId']);
     }
 
+    /**
+     * @throws SorryThatIsAnInvalid
+     */
     public function shepherdId(): ShepherdId
     {
         return ShepherdId::fromString($this->payload['shepherdId']);
