+++ src/Elewant/Herding/Application/Commands/AdoptElePHPantHandler.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Elewant\Herding\Model\Handlers;
+namespace Elewant\Herding\Application\Commands;
 
-use Elewant\Herding\Model\Commands\AdoptElePHPant;
-use Elewant\Herding\Model\HerdCollection;
-use Elewant\Herding\Model\SorryIDoNotHaveThat;
+use Elewant\Herding\DomainModel\Herd\HerdCollection;
+use Elewant\Herding\DomainModel\SorryICanNotChangeHerd;
+use Elewant\Herding\DomainModel\SorryIDoNotHaveThat;
+use Elewant\Herding\DomainModel\SorryThatIsAnInvalid;
 
 final class AdoptElePHPantHandler
 {
@@ -23,9 +24,11 @@ final class AdoptElePHPantHandler
     /**
      * @param AdoptElePHPant $command
      *
-     * @throws SorryIDoNotHaveThat (herd)
+     * @throws SorryICanNotChangeHerd
+     * @throws SorryIDoNotHaveThat
+     * @throws SorryThatIsAnInvalid
      */
-    public function __invoke(AdoptElePHPant $command)
+    public function __invoke(AdoptElePHPant $command): void
     {
         $herd = $this->herdCollection->get($command->herdId());
         if (!$herd) {
