+++ src/Elewant/Herding/Application/Commands/EliminateDesireForBreedHandler.php
@@ -2,11 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Elewant\Herding\Model\Handlers;
+namespace Elewant\Herding\Application\Commands;
 
-use Elewant\Herding\Model\Commands\EliminateDesireForBreed;
-use Elewant\Herding\Model\HerdCollection;
-use Elewant\Herding\Model\SorryIDoNotHaveThat;
+use Elewant\Herding\DomainModel\Herd\HerdCollection;
+use Elewant\Herding\DomainModel\SorryICanNotChangeHerd;
+use Elewant\Herding\DomainModel\SorryIDoNotHaveThat;
+use Elewant\Herding\DomainModel\SorryThatIsAnInvalid;
 
 final class EliminateDesireForBreedHandler
 {
@@ -20,7 +21,14 @@ final class EliminateDesireForBreedHandler
         $this->herdCollection = $herdCollection;
     }
 
-    public function __invoke(EliminateDesireForBreed $command)
+    /**
+     * @param EliminateDesireForBreed $command
+     *
+     * @throws SorryIDoNotHaveThat
+     * @throws SorryICanNotChangeHerd
+     * @throws SorryThatIsAnInvalid
+     */
+    public function __invoke(EliminateDesireForBreed $command): void
     {
         $herd = $this->herdCollection->get($command->herdId());
         if (!$herd) {
